+++ codox.example/src/clojure/codox/example2/core.clj
@@ -1,0 +1,4 @@
+(ns codox.example2.core
+  "Namespace with three parts")
+
+(defn inner [x])