+++ codox.leiningen/src/leiningen/doc.clj
@@ -1,18 +1,25 @@
 (ns leiningen.doc
   (:refer-clojure :exclude [doc])
-  (:use [leinjacker.eval :only (eval-in-project)])
-  (:require [leinjacker.deps :as deps]))
+  (:require [leinjacker.deps :as deps]
+            [leinjacker.eval :as eval]
+            [leiningen.core.project :as project]
+            [clojure.string :as str]))
 
 (defn- get-options [project]
-  (-> project
-      (select-keys [:name :version :description])
-      (merge {:sources ["src"]} ;; Default gets overwritten if set in :codox
-             (get project :codox))))
+  (merge {:sources (:source-paths project ["src"])}
+         (-> project :codox)
+         {:name (str/capitalize (:name project))}
+         (select-keys project [:root :version :description])
+         (-> project :codox :project)))
 
 (defn doc
   "Generate API documentation from source code."
   [project]
-  (eval-in-project
-   (deps/add-if-missing project '[codox/codox.core "0.6.7"])
-   `(codox.main/generate-docs '~(get-options project))
-   `(require 'codox.main)))
+  (let [project (if (get-in project [:profiles :codox])
+                  (project/merge-profiles project [:codox])
+                  project)]
+    (eval/eval-in-project
+     (deps/add-if-missing project '[codox/codox.core "0.8.5"])
+     `(codox.main/generate-docs
+       (update-in '~(get-options project) [:src-uri-mapping] eval))
+     `(require 'codox.main))))